fix: remove the setting page
