+++ src/panels/NoteRecommendationsPanel.tsx
@@ -69,36 +69,34 @@ const WidgetListCurrentGroup: React.FC = () => {
   }
 
   return(
-    <div className='overflow-y-scroll'>
-      <div className='space-y-2'>
-        {
-          recoModeContext.recoMode === 'note' ? 
+    <div className='space-y-2'>
+      {
+        recoModeContext.recoMode === 'note' ? 
 
-            listWidgetsQ.isSuccess ? (
-              !listWidgetsQ.data?.widgets.length ? (
-                <div className='my-4 text-center text-sm text-gray-400'>
+          listWidgetsQ.isSuccess ? (
+            !listWidgetsQ.data?.widgets.length ? (
+              <div className='my-4 text-center text-sm text-gray-400'>
                 You have no widgets for this note
-                </div>
-              ) : (
-                listWidgetsQ.data?.widgets?.map((widget, idx) => (
-                  <WidgetListItem key={`widget-list-${idx}`} widget={widget} />
-                ))
-              )
+              </div>
             ) : (
-              <div className='my-4 text-center text-sm text-gray-400'>
-              Loading your widgets...
-              </div>
-            ) : blockWidgetsArray.length >= 1 ?
-              blockWidgetsArray.map((widget, idx) => (
+              listWidgetsQ.data?.widgets?.map((widget, idx) => (
                 <WidgetListItem key={`widget-list-${idx}`} widget={widget} />
-              )) :
-              <div className='my-4 text-center text-sm text-gray-400'>
+              ))
+            )
+          ) : (
+            <div className='my-4 text-center text-sm text-gray-400'>
+              Loading your widgets...
+            </div>
+          ) : blockWidgetsArray.length >= 1 ?
+            blockWidgetsArray.map((widget, idx) => (
+              <WidgetListItem key={`widget-list-${idx}`} widget={widget} />
+            )) :
+            <div className='my-4 text-center text-sm text-gray-400'>
                 You have no widgets for this block
-              </div>
+            </div>
 
-        }
+      }
         
-      </div>
     </div>
   )
 }
