+++ src/views/profile/ProfileView.tsx
@@ -1,11 +1,13 @@
+import { Menu, Transition } from '@headlessui/react'
 import { ArrowPathIcon, CheckIcon, XMarkIcon } from '@heroicons/react/24/outline'
-import { CodeBracketIcon, ExclamationTriangleIcon, InboxIcon, PencilIcon } from '@heroicons/react/24/solid'
+import { ChevronDownIcon, CodeBracketIcon, ExclamationTriangleIcon, InboxIcon, PencilIcon } from '@heroicons/react/24/solid'
 import { getAnalytics, logEvent } from 'firebase/analytics'
-import React, { useState } from 'react'
+import React, { Fragment, useState } from 'react'
 
 import LoaderIcon from '../../components/icons/LoaderIcon'
 import ViewSkeleton from '../../components/view/ViewSkeleton'
 import { useAuthContext } from '../../contexts/auth'
+import { LangageContext } from '../../contexts/langage'
 import { useDeleteMyAccount, useGetAccount, useRegisterToMobileBeta, useUpdateMyAccount } from '../../hooks/api/accounts'
 import { useGetGroup } from '../../hooks/api/groups'
 import { useAcceptInvite, useDenyInvite, useListInvites } from '../../hooks/api/invites'
@@ -13,7 +15,6 @@ import useClickOutside from '../../hooks/click'
 import { FormatMessage } from '../../i18n/TextComponent'
 import { TOGGLE_DEV_FEATURES } from '../../lib/env'
 import { V1Account, V1GroupInvite } from '../../protorepo/openapi/typescript-axios'
-
 const InviteListItem: React.FC<{ invite: V1GroupInvite }> = (props) => {
   const getGroupQ = useGetGroup({ groupId: props.invite.groupId as string })
   const denyInviteQ = useDenyInvite()
@@ -312,6 +313,86 @@ const ProfileViewFeedbackSection: React.FC = () => {
   )
 }
 
+const ProfileChangeLangage: React.FC = () => {
+  const context = React.useContext(LangageContext)
+  const analytics = getAnalytics()
+  
+  if (!TOGGLE_DEV_FEATURES) {
+    logEvent(analytics, 'page_view', {
+      page_title: 'settings'
+    })
+  }
+  function classNames(...classes) {
+    return classes.filter(Boolean).join(' ')
+  }
+
+  return (
+    <div className='relative mt-4 w-full rounded-md border border-gray-100 bg-gray-50'>
+      {/* Header */}
+      <div className='flex items-center justify-between border-b border-[#efefef] p-5'>
+        <div className='flex items-center'>
+          <CodeBracketIcon className='mr-2 h-5 w-5 text-gray-600' />
+          <p className='text-base font-medium text-gray-600'>              
+            <FormatMessage id='SETTINGS.langage.title' />
+          </p>
+        </div>
+      </div>
+      <div className='grid grid-cols-[40%_60%] p-5'>
+        <div className='relative inline-block text-left'>
+          <Menu as='div' className='relative inline-block text-left'>
+            <div>
+              <Menu.Button className='inline-flex w-full justify-center gap-x-1.5 rounded-md bg-white px-3 py-2 text-sm font-semibold text-gray-900 shadow-sm ring-1 ring-inset ring-gray-300 hover:bg-gray-50'>
+                <FormatMessage id='SETTINGS.langage.options' />
+                <ChevronDownIcon className='-mr-1 h-5 w-5 text-gray-400' aria-hidden='true' />
+              </Menu.Button>
+            </div>
+            <Transition
+              as={Fragment}
+              enter='transition ease-out duration-100'
+              enterFrom='transform opacity-0 scale-95'
+              enterTo='transform opacity-100 scale-100'
+              leave='transition ease-in duration-75'
+              leaveFrom='transform opacity-100 scale-100'
+              leaveTo='transform opacity-0 scale-95'
+            >
+              <Menu.Items className='absolute left-0 z-10 mt-2 w-56 origin-top-right rounded-md bg-white shadow-lg ring-1 ring-black focus:outline-none'>
+                <div className='py-1'>
+                  <Menu.Item>
+                    {({ active }) => (
+                      <a
+                        className={classNames(
+                          active ? 'bg-gray-100 text-gray-900' : 'text-gray-700',
+                          'block px-4 py-2 text-sm'
+                        )}
+                        onClick={() => {context?.changeLangage('fr')}}
+                      >
+                        <FormatMessage id='SETTINGS.langage.french' />
+                      </a>
+                    )}
+                  </Menu.Item>
+                  <Menu.Item>
+                    {({ active }) => (
+                      <a
+                        className={classNames(
+                          active ? 'bg-gray-100 text-gray-900' : 'text-gray-700',
+                          'block px-4 py-2 text-sm'
+                        )}
+                        onClick={() => {context?.changeLangage('en')}}
+                      >
+                        <FormatMessage id='SETTINGS.langage.english' />
+                      </a>
+                    )}
+                  </Menu.Item>
+                </div>
+              </Menu.Items>
+            </Transition>
+          </Menu>
+        </div>
+      </div>
+    </div>
+  )
+}
+
 
 const ProfileView: React.FC = () => {
   return (
@@ -326,6 +407,7 @@ const ProfileView: React.FC = () => {
         <hr className='m-5 rounded border-2'></hr>
         <ProfileViewFeedbackSection />
         <hr className='m-5 rounded border-2'></hr>
+        <ProfileChangeLangage />
       </div>
     </ViewSkeleton>
   )
