+++ src/views/signin/SigninView.tsx
@@ -102,7 +102,6 @@ const SigninView: React.FC = () => {
           />
           <OldInput
             label={formatMessage({ id: 'AUTH.pwd' })}
-            type='password'
             value={password}
             onChange={(e) => {
               setPassword(e.target.value)
